河山纷裂终须入掌 百万明珠抬手赐名将
