+++ HMCLCore/src/main/java/org/jackhuang/hmcl/download/forge/ForgeNewInstallProfile.java
@@ -29,47 +29,33 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 @Immutable
-public class ForgeNewInstallProfile implements Validation {
-
-    private final int spec;
-    private final String minecraft;
-    private final String json;
-    private final String version;
-    private final Artifact path;
-    private final List<Library> libraries;
-    private final List<Processor> processors;
-    private final Map<String, Datum> data;
-
-    public ForgeNewInstallProfile(int spec, String minecraft, String json, String version, Artifact path, List<Library> libraries, List<Processor> processors, Map<String, Datum> data) {
-        this.spec = spec;
-        this.minecraft = minecraft;
-        this.json = json;
-        this.version = version;
-        this.path = path;
-        this.libraries = libraries;
-        this.processors = processors;
-        this.data = data;
-    }
+public record ForgeNewInstallProfile(int spec, String minecraft, String json, String version, Artifact path,
+                                     List<Library> libraries, List<Processor> processors,
+                                     Map<String, Datum> data) implements Validation {
 
     /**
      * Specification for install_profile.json.
      */
-    public int getSpec() {
+    @Override
+    public int spec() {
         return spec;
     }
 
     /**
      * Vanilla game version that this installer supports.
      */
-    public String getMinecraft() {
+    @Override
+    public String minecraft() {
         return minecraft;
     }
 
     /**
      * Version json to be installed.
+     *
      * @return path of the version json relative to the installer JAR file.
      */
-    public String getJson() {
+    @Override
+    public String json() {
         return json;
     }
 
@@ -77,12 +63,14 @@ public class ForgeNewInstallProfile implements Validation {
      *
      * @return forge version.
      */
-    public String getVersion() {
+    @Override
+    public String version() {
         return version;
     }
 
     /**
      * Maven artifact path for the main jar to install.
+     *
      * @return artifact path of the main jar.
      */
     public Optional<Artifact> getPath() {
@@ -91,16 +79,19 @@ public class ForgeNewInstallProfile implements Validation {
 
     /**
      * Libraries that processors depend on.
+     *
      * @return the required dependencies.
      */
-    public List<Library> getLibraries() {
+    @Override
+    public List<Library> libraries() {
         return libraries == null ? Collections.emptyList() : libraries;
     }
 
     /**
      * Tasks to be executed to setup modded environment.
      */
-    public List<Processor> getProcessors() {
+    @Override
+    public List<Processor> processors() {
         if (processors == null) return Collections.emptyList();
         return processors.stream().filter(p -> p.isSide("client")).collect(Collectors.toList());
     }
@@ -114,7 +105,7 @@ public class ForgeNewInstallProfile implements Validation {
         if (data == null)
             return new HashMap<>();
 
-        return data.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().getClient()));
+        return data.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().client()));
     }
 
     @Override
@@ -123,23 +114,12 @@ public class ForgeNewInstallProfile implements Validation {
             throw new JsonParseException("ForgeNewInstallProfile is malformed");
     }
 
-    public static class Processor implements Validation {
-        private final List<String> sides;
-        private final Artifact jar;
-        private final List<Artifact> classpath;
-        private final List<String> args;
-        private final Map<String, String> outputs;
-
-        public Processor(List<String> sides, Artifact jar, List<Artifact> classpath, List<String> args, Map<String, String> outputs) {
-            this.sides = sides;
-            this.jar = jar;
-            this.classpath = classpath;
-            this.args = args;
-            this.outputs = outputs;
-        }
+    public record Processor(List<String> sides, Artifact jar, List<Artifact> classpath, List<String> args,
+                            Map<String, String> outputs) implements Validation {
 
         /**
          * Check which side this processor should be run on. We only support client install currently.
+         *
          * @param side can be one of "client", "server", "extract".
          * @return true if the processor can run on the side.
          */
@@ -149,17 +129,21 @@ public class ForgeNewInstallProfile implements Validation {
 
         /**
          * The executable jar of this processor task. Will be executed in installation process.
+         *
          * @return the artifact path of executable jar.
          */
-        public Artifact getJar() {
+        @Override
+        public Artifact jar() {
             return jar;
         }
 
         /**
          * The dependencies of this processor task.
+         *
          * @return the artifact path of dependencies.
          */
-        public List<Artifact> getClasspath() {
+        @Override
+        public List<Artifact> classpath() {
             return classpath == null ? Collections.emptyList() : classpath;
         }
 
@@ -170,10 +154,12 @@ public class ForgeNewInstallProfile implements Validation {
          * {entry}: Get corresponding value of the entry in {@link ForgeNewInstallProfile#getData()}
          * {MINECRAFT_JAR}: path of the Minecraft jar.
          * {SIDE}: values other than "client" will be ignored.
+         *
          * @return arguments to pass to the processor jar.
          * @see ForgeNewInstallTask#parseLiteral(String, Map, ExceptionalFunction)
          */
-        public List<String> getArgs() {
+        @Override
+        public List<String> args() {
             return args == null ? Collections.emptyList() : args;
         }
 
@@ -182,10 +168,12 @@ public class ForgeNewInstallProfile implements Validation {
          * Arguments to pass to the processor jar.
          * Keys can be in one of [artifact] or {entry}. Should be file path.
          * Values can be in one of {entry} or 'literal'. Should be SHA-1 checksum.
+         *
          * @return files output from this processor.
          * @see ForgeNewInstallTask#parseLiteral(String, Map, ExceptionalFunction)
          */
-        public Map<String, String> getOutputs() {
+        @Override
+        public Map<String, String> outputs() {
             return outputs == null ? Collections.emptyMap() : outputs;
         }
 
@@ -196,21 +184,18 @@ public class ForgeNewInstallProfile implements Validation {
         }
     }
 
-    public static class Datum {
-        private final String client;
-
-        public Datum(String client) {
-            this.client = client;
-        }
+    public record Datum(String client) {
 
         /**
          * Can be in the following formats:
          * [value]: An artifact path.
          * 'value': A string literal.
          * value: A file in the installer package, to be extracted to a temp folder, and then have the absolute path in replacements.
+         *
          * @return Value to use for the client install
          */
-        public String getClient() {
+        @Override
+        public String client() {
             return client;
         }
     }
