+++ HMCLCore/src/main/java/org/jackhuang/hmcl/download/neoforge/NeoForgeOldInstallTask.java
@@ -21,9 +21,10 @@ import org.jackhuang.hmcl.download.ArtifactMalformedException;
 import org.jackhuang.hmcl.download.DefaultDependencyManager;
 import org.jackhuang.hmcl.download.forge.ForgeNewInstallProfile;
 import org.jackhuang.hmcl.download.forge.ForgeNewInstallProfile.Processor;
-import org.jackhuang.hmcl.download.game.GameLibrariesTask;
 import org.jackhuang.hmcl.download.game.GameInstanceJsonDownloadTask;
+import org.jackhuang.hmcl.download.game.GameLibrariesTask;
 import org.jackhuang.hmcl.game.*;
+import org.jackhuang.hmcl.java.JavaRuntime;
 import org.jackhuang.hmcl.task.FileDownloadTask;
 import org.jackhuang.hmcl.task.Task;
 import org.jackhuang.hmcl.util.DigestUtils;
@@ -34,7 +35,6 @@ import org.jackhuang.hmcl.util.io.ChecksumMismatchException;
 import org.jackhuang.hmcl.util.io.CompressingUtils;
 import org.jackhuang.hmcl.util.io.FileUtils;
 import org.jackhuang.hmcl.util.platform.CommandBuilder;
-import org.jackhuang.hmcl.java.JavaRuntime;
 import org.jackhuang.hmcl.util.platform.SystemUtils;
 import org.jetbrains.annotations.NotNull;
 
@@ -53,8 +53,8 @@ import java.util.jar.Attributes;
 import java.util.jar.JarFile;
 import java.util.zip.ZipException;
 
-import static org.jackhuang.hmcl.util.logging.Logger.LOG;
 import static org.jackhuang.hmcl.util.gson.JsonUtils.fromNonNullJson;
+import static org.jackhuang.hmcl.util.logging.Logger.LOG;
 
 public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
 
@@ -74,7 +74,7 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
             Map<String, String> outputs = new HashMap<>();
             boolean miss = false;
 
-            for (Map.Entry<String, String> entry : processor.getOutputs().entrySet()) {
+            for (Map.Entry<String, String> entry : processor.outputs().entrySet()) {
                 String key = entry.getKey();
                 String value = entry.getValue();
 
@@ -105,11 +105,11 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
                 }
             }
 
-            if (!processor.getOutputs().isEmpty() && !miss) {
+            if (!processor.outputs().isEmpty() && !miss) {
                 return;
             }
 
-            Path jar = gameRepository.getLayout().getArtifactFile(processor.getJar());
+            Path jar = gameRepository.getLayout().getArtifactFile(processor.jar());
             if (!Files.isRegularFile(jar))
                 throw new FileNotFoundException("Game processor file not found, should be downloaded in preprocess");
 
@@ -125,8 +125,8 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
             command.add(JavaRuntime.getDefault().getBinary().toString());
             command.add("-cp");
 
-            List<String> classpath = new ArrayList<>(processor.getClasspath().size() + 1);
-            for (Artifact artifact : processor.getClasspath()) {
+            List<String> classpath = new ArrayList<>(processor.classpath().size() + 1);
+            for (Artifact artifact : processor.classpath()) {
                 Path file = gameRepository.getLayout().getArtifactFile(artifact);
                 if (!Files.isRegularFile(file))
                     throw new Exception("Game processor dependency missing");
@@ -137,8 +137,8 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
 
             command.add(mainClass);
 
-            List<String> args = new ArrayList<>(processor.getArgs().size());
-            for (String arg : processor.getArgs()) {
+            List<String> args = new ArrayList<>(processor.args().size());
+            for (String arg : processor.args()) {
                 String parsed = parseLiteral(arg, vars);
                 if (parsed == null)
                     throw new ArtifactMalformedException("Invalid forge installation configuration");
@@ -147,7 +147,7 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
 
             command.addAll(args);
 
-            LOG.info("Executing external processor " + processor.getJar().toString() + ", command line: " + new CommandBuilder().addAll(command).toString());
+            LOG.info("Executing external processor " + processor.jar().toString() + ", command line: " + new CommandBuilder().addAll(command).toString());
             int exitCode = SystemUtils.callExternalProcess(command);
             if (exitCode != 0)
                 throw new IOException("Game processor exited abnormally with code " + exitCode);
@@ -288,10 +288,10 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
     public void preExecute() throws Exception {
         try (FileSystem fs = CompressingUtils.createReadOnlyZipFileSystem(installer)) {
             profile = JsonUtils.fromNonNullJson(Files.readString(fs.getPath("install_profile.json")), ForgeNewInstallProfile.class);
-            processors = profile.getProcessors();
-            neoForgeVersion = JsonUtils.fromNonNullJson(Files.readString(fs.getPath(profile.getJson())), GameInstanceManifest.class);
+            processors = profile.processors();
+            neoForgeVersion = JsonUtils.fromNonNullJson(Files.readString(fs.getPath(profile.json())), GameInstanceManifest.class);
 
-            for (Library library : profile.getLibraries()) {
+            for (Library library : profile.libraries()) {
                 Path file = fs.getPath("maven").resolve(library.getPath());
                 if (Files.exists(file)) {
                     Path dest = gameRepository.getLayout().getLibraryFile(manifest.id(), library);
@@ -310,7 +310,7 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
             throw new ArtifactMalformedException("Malformed forge installer file", ex);
         }
 
-        dependents.add(new GameLibrariesTask(dependencyManager, manifest, true, profile.getLibraries()));
+        dependents.add(new GameLibrariesTask(dependencyManager, manifest, true, profile.libraries()));
     }
 
     private Map<String, String> parseOptions(List<String> args, Map<String, String> vars) {
@@ -338,7 +338,7 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
     }
 
     private Task<?> patchDownloadMojangMappingsTask(Processor processor, Map<String, String> vars) {
-        Map<String, String> options = parseOptions(processor.getArgs(), vars);
+        Map<String, String> options = parseOptions(processor.args(), vars);
         if (!"DOWNLOAD_MOJMAPS".equals(options.get("task")) || !"client".equals(options.get("side")))
             return null;
         String version = options.get("version");
@@ -408,7 +408,7 @@ public class NeoForgeOldInstallTask extends Task<GameInstancePatch> {
 
         vars.put("SIDE", "client");
         vars.put("MINECRAFT_JAR", FileUtils.getAbsolutePath(isolatedMinecraftJar));
-        vars.put("MINECRAFT_VERSION", profile.getMinecraft());
+        vars.put("MINECRAFT_VERSION", profile.minecraft());
         vars.put("ROOT", FileUtils.getAbsolutePath(gameRepository.getBaseDirectory()));
         vars.put("INSTALLER", installer.toAbsolutePath().toString());
         vars.put("LIBRARY_DIR", FileUtils.getAbsolutePath(gameRepository.getLayout().getLibrariesDirectory()));
